requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-api-python-client>=2.0.0
cohere>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
dash>=2.14.0
cryptography>=41.0.0
PyJWT>=2.8.0
python-jose>=3.3.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
PyQt5>=5.15.0
spacy>=3.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
